refactor: сервисы ItemService и UserService реализованы как интерфейс, их имплементации вынесены в отдельные файлы; в контроллере ItemController название заголовка вынесено в константу
